Split the AgentHealthDetailsBar out into its own file
Eventually I would like to move all the widgets into their own directory
and all the data objects into their own directory. This is one of the few
files that has both in one file.
