+++ app_flutter/test/agent_health_details_bar_test.dart
@@ -1,0 +1,94 @@
+// Copyright (c) 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:fixnum/fixnum.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:cocoon_service/protos.dart' show Agent;
+
+import 'package:app_flutter/agent_health_details.dart';
+import 'package:app_flutter/agent_health_details_bar.dart';
+
+void main() {
+  testWidgets('healthy bar', (WidgetTester tester) async {
+    final Agent agent = Agent()
+      ..healthCheckTimestamp = Int64.parseInt(DateTime.now().millisecondsSinceEpoch.toString())
+      ..isHealthy = true
+      ..healthDetails = '''
+ssh-connectivity: succeeded
+  Last known IP address: 192.168.1.29
+
+android-device-ZY223D6B7B: succeeded
+has-healthy-devices: succeeded
+  Found 1 healthy devices
+
+cocoon-authentication: succeeded
+cocoon-connection: succeeded
+able-to-perform-health-check: succeeded''';
+
+    final AgentHealthDetails healthDetails = AgentHealthDetails(agent);
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: AgentHealthDetailsBar(healthDetails),
+      ),
+    );
+
+    expect(find.byIcon(Icons.timer), findsNothing);
+    expect(find.byIcon(Icons.verified_user), findsOneWidget);
+    expect(find.byIcon(Icons.network_wifi), findsOneWidget);
+    expect(find.byIcon(Icons.devices), findsOneWidget);
+  });
+
+  testWidgets('timed out icon', (WidgetTester tester) async {
+    final Agent agent = Agent()
+      ..healthCheckTimestamp = Int64.parseInt('100')
+      ..isHealthy = true
+      ..healthDetails = '''
+ssh-connectivity: succeeded
+  Last known IP address: 192.168.1.29
+
+android-device-ZY223D6B7B: succeeded
+has-healthy-devices: succeeded
+  Found 1 healthy devices
+
+cocoon-authentication: succeeded
+cocoon-connection: succeeded
+able-to-perform-health-check: succeeded''';
+
+    final AgentHealthDetails healthDetails = AgentHealthDetails(agent);
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: AgentHealthDetailsBar(healthDetails),
+      ),
+    );
+
+    expect(find.byIcon(Icons.timer), findsOneWidget);
+    expect(find.byIcon(Icons.verified_user), findsOneWidget);
+    expect(find.byIcon(Icons.network_wifi), findsOneWidget);
+    expect(find.byIcon(Icons.devices), findsOneWidget);
+  });
+
+  testWidgets('unhealthy bar', (WidgetTester tester) async {
+    final Agent agent = Agent()
+      ..healthCheckTimestamp = Int64.parseInt('100')
+      ..isHealthy = false
+      ..healthDetails = '';
+
+    final AgentHealthDetails healthDetails = AgentHealthDetails(agent);
+
+    await tester.pumpWidget(
+      MaterialApp(
+        home: AgentHealthDetailsBar(healthDetails),
+      ),
+    );
+
+    expect(find.byIcon(Icons.timer), findsOneWidget);
+    expect(find.byIcon(Icons.verified_user), findsNothing);
+    expect(find.byIcon(Icons.network_wifi), findsNothing);
+    expect(find.byIcon(Icons.devices), findsNothing);
+  });
+}
