+++ app_flutter/test/agent_tile_test.dart
@@ -10,6 +10,7 @@ import 'package:mockito/mockito.dart';
 import 'package:cocoon_service/protos.dart' show Agent;
 
 import 'package:app_flutter/agent_health_details.dart';
+import 'package:app_flutter/agent_health_details_bar.dart';
 import 'package:app_flutter/agent_list.dart';
 import 'package:app_flutter/agent_tile.dart';
